Move GUI merge code from sample.c to guimerge.c
