+++ src/sample.h
@@ -98,6 +98,5 @@ void stop_sample();
 int sample_open_file(const char *, GraphData *, double *);
 void sample_close_file();
 void sample_write_files(GList *, WriteInfo *, char *);
-void sample_merge_files(char *merge_filename, GList *filenames, WriteInfo *write_info);
 
 #endif /* SAMPLE_H*/
